// src/api.d.ts
export interface components {
    schemas: {
// ...
        "Model.Ref": {
            id: string;
            providerID: string;
            variant?: string;
        };
// ...
        "Model.Info": {
            id: string;
            modelID: string;
            providerID: string;
            canonical?: string;
            family?: string;
            name: string;
            compatibility?: components["schemas"]["Model.Compatibility"];
            package?: string;
            settings?: components["schemas"]["Model.Settings"];
            headers?: {
                [key: string]: string;
            };
            body?: Record<string, never>;
            capabilities: components["schemas"]["Model.Capabilities"];
            variants: components["schemas"]["Model.Variant"][];
            time: {
                released: number;
            };
            cost: components["schemas"]["Model.Cost"][];
            /** @enum {string} */
            status: "alpha" | "beta" | "deprecated" | "active";
            enabled: boolean;
            limit: {
                context: number;
                input?: number;
                output: number;
            };
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}